Dharma-Platform-Live-Dashboard-master/services/event-bus-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
temporalio==1.5.1
pydantic-settings==2.1.0
pydantic==2.5.0
asyncio-mqtt==0.13.0
redis==5.0.1
python-multipart==0.0.6